# apps/gkx-web-app/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/goalkeepers": {
      "post": {
        "operationId": "GoalkeepersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalkeeperDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Goalkeeper profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper profile created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GoalkeeperModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Crear perfil de portero",
        "tags": [
          "Goalkeepers"
        ]
      },
      "get": {
        "operationId": "GoalkeepersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Goalkeeper profiles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper profiles retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GoalkeeperModel"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Listar perfiles de portero",
        "tags": [
          "Goalkeepers"
        ]
      }
    },
    "/api/goalkeepers/{id}": {
      "get": {
        "operationId": "GoalkeepersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del portero",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goalkeeper profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper profile retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GoalkeeperModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Obtener perfil de portero por id",
        "tags": [
          "Goalkeepers"
        ]
      },
      "patch": {
        "operationId": "GoalkeepersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del portero",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalkeeperDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goalkeeper profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper profile updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GoalkeeperModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Actualizar perfil de portero",
        "tags": [
          "Goalkeepers"
        ]
      },
      "delete": {
        "operationId": "GoalkeepersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del portero",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goalkeeper profile deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper profile deleted successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeleteDataModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Eliminar perfil de portero",
        "tags": [
          "Goalkeepers"
        ]
      }
    },
    "/api/goalkeepers/{id}/progress": {
      "get": {
        "operationId": "GoalkeepersController_getProgress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del portero",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goalkeeper progress retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper progress retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GoalkeeperProgressModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Obtener progreso agregado del portero",
        "tags": [
          "Goalkeepers"
        ]
      }
    },
    "/api/goalkeepers/{id}/evaluations": {
      "get": {
        "operationId": "GoalkeepersController_getEvaluations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del portero",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Goalkeeper evaluations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper evaluations retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EvaluationModel"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Listar evaluaciones del portero",
        "tags": [
          "Goalkeepers"
        ]
      }
    },
...
    "/api/teams/{id}/goalkeepers/{goalkeeperId}": {
      "post": {
        "operationId": "TeamsController_assignGoalkeeper",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identificador del equipo",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "goalkeeperId",
            "required": true,
            "in": "path",
            "description": "Identificador del portero",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Goalkeeper assigned to team successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Goalkeeper assigned to team successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/GoalkeeperModel"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "401": {
            "description": "No autenticado o token invalido"
          },
          "403": {
            "description": "No tiene permisos para esta accion"
          },
          "404": {
            "description": "Recurso no encontrado"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Asignar portero a equipo",
        "tags": [
          "Teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "GKX API",
    "description": "Documentacion de la API GKX",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT en formato Bearer"
      }
    },
    "schemas": {
...
      "DeleteDataModel": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "GoalkeeperModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "dateOfBirth": {
            "type": "object",
            "nullable": true
          },
          "dominantHand": {
            "type": "object",
            "nullable": true
          },
          "dominantFoot": {
            "type": "object",
            "nullable": true
          },
          "height": {
            "type": "object",
            "nullable": true
          },
          "weight": {
            "type": "object",
            "nullable": true
          },
          "category": {
            "type": "object",
            "nullable": true
          },
          "teamId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "medicalNotes": {
            "type": "object",
            "nullable": true
          },
          "parentContact": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "userId",
          "dateOfBirth",
          "dominantHand",
          "dominantFoot",
          "height",
          "weight",
          "category",
          "teamId",
          "medicalNotes",
          "parentContact",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateGoalkeeperDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "2008-05-24"
          },
          "dominantHand": {
            "type": "string",
            "example": "RIGHT"
          },
          "dominantFoot": {
            "type": "string",
            "example": "RIGHT"
          },
          "height": {
            "type": "number",
            "example": 1.82
          },
          "weight": {
            "type": "number",
            "example": 74.5
          },
          "category": {
            "type": "string",
            "example": "U17"
          },
          "teamId": {
            "type": "string",
            "format": "uuid"
          },
          "medicalNotes": {
            "type": "string",
            "example": "Lesion previa en hombro izquierdo"
          },
          "parentContact": {
            "type": "string",
            "example": "+59170000000"
          }
        },
        "required": [
          "tenantId",
          "userId"
        ]
      },
      "GoalkeeperProgressModel": {
        "type": "object",
        "properties": {
          "evaluationsCount": {
            "type": "number"
          },
          "averageOverallScore": {
            "type": "object",
            "nullable": true
          },
          "latestEvaluationDate": {
            "type": "object",
            "nullable": true
          },
          "latestOverallScore": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "evaluationsCount",
          "averageOverallScore",
          "latestEvaluationDate",
          "latestOverallScore"
        ]
      },
      "EvaluationModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "goalkeeperId": {
            "type": "string",
            "format": "uuid"
          },
          "coachId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "handling": {
            "type": "number"
          },
          "diving": {
            "type": "number"
          },
          "positioning": {
            "type": "number"
          },
          "reflexes": {
            "type": "number"
          },
          "communication": {
            "type": "number"
          },
          "footwork": {
            "type": "number"
          },
          "distribution": {
            "type": "number"
          },
          "aerialPlay": {
            "type": "number"
          },
          "oneVsOne": {
            "type": "number"
          },
          "mentality": {
            "type": "number"
          },
          "overallScore": {
            "type": "number"
          },
          "comments": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "goalkeeperId",
          "coachId",
          "date",
          "handling",
          "diving",
          "positioning",
          "reflexes",
          "communication",
          "footwork",
          "distribution",
          "aerialPlay",
          "oneVsOne",
          "mentality",
          "overallScore",
          "comments",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateGoalkeeperDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}